refactor module
