+++ go.mod
@@ -1,15 +1,10 @@
-module github.com/ch3ck/youtube-dl
+module youtube-dl
 
-go 1.12
+go 1.17
 
 require (
-	github.com/konsorten/go-windows-terminal-sequences v1.0.2 // indirect
-	github.com/sirupsen/logrus v1.5.0
-	github.com/stretchr/objx v0.2.0 // indirect
-	github.com/stretchr/testify v1.4.0 // indirect
-	github.com/wader/goutubedl v0.0.0-20200327095909-c841a70bbbad
-	github.com/wader/osleaktest v0.0.0-20191111175233-f643b0fed071 // indirect
-	golang.org/x/sys v0.0.0-20200408040146-ea54a3c99b9b // indirect
-	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
-	gopkg.in/yaml.v2 v2.2.7 // indirect
+	github.com/sirupsen/logrus v1.8.1
+	github.com/wader/goutubedl v0.0.0-20211221155015-039073da1ac8
 )
+
+require golang.org/x/sys v0.0.0-20191026070338-33540a1f6037 // indirect
