+++ pkg/downloader/download_test.go
@@ -1,4 +1,4 @@
-package main
+package downloader
 
 import (
 	"testing"
@@ -22,7 +22,7 @@ func TestApi(t *testing.T) {
 
 	// path := "test"
 	for i, table := range tables {
-		err := decodeVideoStream(table.url, "mp3")
+		err := DecodeVideoStream(table.url, "mp3")
 		if err != nil {
 			t.Errorf("videoId(%d): expected %q, actual %q", i, table.id, err)
 		}
@@ -31,7 +31,7 @@ func TestApi(t *testing.T) {
 
 func BenchmarkVideoId(b *testing.B) {
 	for n := 0; n < b.N; n++ {
-		if err := decodeVideoStream(tables[0].url, "mp3"); err != nil {
+		if err := DecodeVideoStream(tables[0].url, "mp3"); err != nil {
 			b.Errorf("Error downloading video: %v", err)
 		}
 	}
